+++ LCD.cpp
@@ -1,6 +1,6 @@
 #include "LCD.h"
 
-void LCD::drawVolts(float v, float max, float min, int x, int y, int w, int h, int b) {
+void LCD::drawVolts(float v, float max, float min, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t b) {
   float level = (v - min) / (max - min);
   if (level > 1) level = 1;
   if (level < 0) level = 0;
@@ -10,17 +10,17 @@ void LCD::drawVolts(float v, float max, float min, int x, int y, int w, int h, i
   display.setCursor(x+w+5, y+2);
 }
 
-void LCD::drawSignal(float v, float max, float min, int x, int y, int w, int h, int b) {
+void LCD::drawSignal(float v, float max, float min, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t b) {
   float level = (v - min) / (max - min);
   if (level > 1) level = 1;
   if (level < 0) level = 0;
-  int numBars = (int) (level * 6 + 0.5);
-  int heightIncrement = h / 5;
-  int barWidth = w/5 - b;
+  uint8_t numBars = (uint8_t) (level * 6 + 0.5);
+  uint8_t heightIncrement = h / 5;
+  uint8_t barWidth = w/5 - b;
 
-  int cur = x;
+  uint8_t cur = x;
 
-  for (int i = 0; i < numBars; ++i) {
+  for (uint8_t i = 0; i < numBars; ++i) {
     display.fillRect(cur, y+(4-i)*heightIncrement, barWidth, heightIncrement*(i+1),  WHITE);
     cur += barWidth + b;
   }
@@ -31,7 +31,7 @@ void LCD::updateHeaders() {
   drawVolts(this->volts, MAX_VOLTS, MIN_VOLTS, 0, 5, 18, 10, 1);
   display.print(this->modelName);
 
-  int i = 0;
+  uint8_t i = 0;
 
   drawSignal(rssi, 100, 0, display.width() - 18, 5, 18, 10, 1);
   if (a1 != 0) {
@@ -45,28 +45,28 @@ void LCD::updateHeaders() {
 }
 
 void LCD::updateInputs() {
-  int width = (display.width()/2 - (numInputs - 1) * SPACE - SEPARATOR) / numInputs;
-  int height = (display.height() - 24)/2;
+  uint8_t width = (display.width()/2 - (numInputs - 1) * SPACE - SEPARATOR) / numInputs;
+  uint8_t height = (display.height() - 24)/2;
 
-  for (int i = 0; i < numInputs; ++i) {
-    int amp = (inputs[i]+1)*height +1;
+  for (uint8_t i = 0; i < numInputs; ++i) {
+    uint8_t amp = (inputs[i]+1)*height +1;
     display.fillRect(i*SPACE+i*width, display.height() - amp, width, amp, 1);
   }
 }
 
 void LCD::updateChannels() {
-  int width = (display.width()/2 - (numChannels - 1) * SPACE - SEPARATOR) / numChannels;
-  int height = (display.height() - 24)/2;
-  int offset = display.width()/2 + SEPARATOR;
+  uint8_t width = (display.width()/2 - (numChannels - 1) * SPACE - SEPARATOR) / numChannels;
+  uint8_t height = (display.height() - 24)/2;
+  uint8_t offset = display.width()/2 + SEPARATOR;
 
-  for (int i = 0; i < numChannels; ++i) {
-    int amp = (channels[i]+1)*height +1;
+  for (uint8_t i = 0; i < numChannels; ++i) {
+    uint8_t amp = (channels[i]+1)*height +1;
     display.fillRect(offset + i*SPACE + i*width, display.height() - amp, width, amp, 1);
   }
 }
 
 void LCD::setModelName(char *name) {
-  int i = 0;
+  uint8_t i = 0;
   for (; i < LENGTH_NAME; ++i) {
     if (name[i] == '\0') {
       break;
@@ -78,12 +78,12 @@ void LCD::setModelName(char *name) {
   }
 }
 
-void LCD::setInputs(float *inputs, int n) {
+void LCD::setInputs(float *inputs, uint8_t n) {
   this->inputs = inputs;
   this->numInputs = n;
 }
 
-void LCD::setChannels(float *channels, int n) {
+void LCD::setChannels(float *channels, uint8_t n) {
   this->channels = channels;
   this->numChannels = n;
 }
