+++ LCD.h
@@ -24,20 +24,20 @@ class LCD {
  private:
 
   float volts;
-  int rssi;
-  int a1;
-  int a2;
+  uint8_t rssi;
+  uint8_t a1;
+  uint8_t a2;
   char modelName[LENGTH_NAME+1] = {};
   float *inputs;
   float *channels;
-  int numInputs;
-  int numChannels;
+  uint8_t numInputs;
+  uint8_t numChannels;
   Adafruit_SSD1306 display;
 
  private:
 
-  void drawVolts(float v, float max, float min, int x, int y, int w, int h, int b);
-  void drawSignal(float v, float max, float min, int x, int y, int w, int h, int b);
+  void drawVolts(float v, float max, float min, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t b);
+  void drawSignal(float v, float max, float min, uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t b);
   void updateHeaders();
   void updateInputs();
   void updateChannels();
@@ -48,12 +48,12 @@ class LCD {
 
   inline void begin();
   inline void setVolts(float volts);
-  inline void setRssi(int rssi);
-  inline void setA1(int a1);
-  inline void setA2(int a2);
+  inline void setRssi(uint8_t rssi);
+  inline void setA1(uint8_t a1);
+  inline void setA2(uint8_t a2);
   void setModelName(char *name);
-  void setInputs(float *inputs, int n);
-  void setChannels(float *channels, int n);
+  void setInputs(float *inputs, uint8_t n);
+  void setChannels(float *channels, uint8_t n);
   void update();
 };
 
@@ -68,15 +68,15 @@ inline void LCD::setVolts(float volts) {
   this->volts = volts;
 }
 
-inline void LCD::setRssi(int rssi) {
+inline void LCD::setRssi(uint8_t rssi) {
   this->rssi = rssi;
 }
 
-inline void LCD::setA1(int a1) {
+inline void LCD::setA1(uint8_t a1) {
   this->a1 = a1;
 }
 
-inline void LCD::setA2(int a2) {
+inline void LCD::setA2(uint8_t a2) {
   this->a2 = a2;
 }
