+++ Radio.h
@@ -15,9 +15,9 @@ typedef struct model_ {
   float *drLow;
   float *drHigh;
   float *transform;
-  int numChannels;
+  uint8_t numChannels;
   float *channels;
-  int numInputs;
+  uint8_t numInputs;
   float *inputs;
   float *trim;
 } Model;
@@ -30,17 +30,17 @@ class Radio {
   float *channels;
   float *trim;
   bool lowRates;
-  int pin;
+  uint8_t pin;
 
  public:
 
-  Radio(int pin);
+  Radio(uint8_t pin);
   void begin();
   void setModel(Model *m);
   void update(float *inputs);
   float *getChannels();
   float *getTrim();
-  void setTrim(float *trim, int n);
+  void setTrim(float *trim, uint8_t n);
   void setLowRates();
   void setHighRates();
 
@@ -48,11 +48,11 @@ class Radio {
 
   float toServo(float *input);
   inline float expo(float a, float x);
-  void mix(float *transform, float *inputs, float *channels, int n, int m);
-  inline void init(float *x, int n, float c);
-  inline void add(float *y, float *x1, float *x2, int n);
-  inline void scale(float *y, float *x, int scale, int n);
-  inline void mult(float *y, float *x1, float *x2, int n);
+  void mix(float *transform, float *inputs, float *channels, uint8_t n, uint8_t m);
+  inline void init(float *x, uint8_t n, float c);
+  inline void add(float *y, float *x1, float *x2, uint8_t n);
+  inline void scale(float *y, float *x, uint8_t scale, uint8_t n);
+  inline void mult(float *y, float *x1, float *x2, uint8_t n);
 };
 
 // y = a*x^3+(1-a)x, a .. expo (0 >= a >= 1), x .. input (-1 <= x <= 1)
@@ -60,26 +60,26 @@ inline float Radio::expo(float a, float x) {
   return a*x*x*x+(1-a)*x;
 }
 
-inline void Radio::init(float *x, int n, float c) {
-  for (int i = 0; i < n; ++i) {
+inline void Radio::init(float *x, uint8_t n, float c) {
+  for (uint8_t i = 0; i < n; ++i) {
     x[i] = c;
   }
 }
 
-inline void Radio::add(float *y, float *x1, float *x2, int n) {
-  for (int i = 0; i < n; ++i) {
+inline void Radio::add(float *y, float *x1, float *x2, uint8_t n) {
+  for (uint8_t i = 0; i < n; ++i) {
     y[i] = x1[i] + x2[i];
   }
 }
 
-inline void Radio::scale(float *y, float *x, int scale, int n) {
-  for (int i = 0; i < n; ++i) {
+inline void Radio::scale(float *y, float *x, uint8_t scale, uint8_t n) {
+  for (uint8_t i = 0; i < n; ++i) {
     y[i] = scale * x[i];
   }
 }
 
-inline void Radio::mult(float *y, float *x1, float *x2, int n) {
-  for (int i = 0; i < n; ++i) {
+inline void Radio::mult(float *y, float *x1, float *x2, uint8_t n) {
+  for (uint8_t i = 0; i < n; ++i) {
     y[i] = x1[i] * x2[i];
   }
 }
